Move error reporting into context
By moving the error pointer into context and initializing it to NULL, we
get an easy check for whether or not we have encountered an error.
Rather than needing to check the result of each individual operation, we
can now collect a few values (as in inference.c) and check if any of
them failed afterward.

This makes the code more straight-forward and easy to read.