First crack at iterative analyze
The code in this commit does work, in that it produces the correct
typing for our example program, factorial. However, I think the approach
is wrong. In particular I don't think it can contend with local scope or
shadowing of higher level variables generally, this is because of the
opportunistic resolution in IDENTIFIER.

I have an alternative approach in mind which only attempts to 'resolve'
down concrete types at lambda, let, and letrec terms, as those represent
boundaries between global scope and local scope.